+++ .github/workflows/backend.yml
@@ -21,7 +21,7 @@ jobs:
           python -m pip install --upgrade pip
           pip install --user pipenv
           pipenv --python ${{ matrix.python-version }}
-          cd machine-learning-client
+          cd machine_learning_client
           pipenv install --dev
           pipenv lock
           pipenv install -r requirements.txt
@@ -32,5 +32,10 @@ jobs:
 
       - name: Testing with pytest
         run: |
-          cd machine-learning-client/tests
+          cd machine_learning_client/tests
           pipenv run python -m pytest
+      
+      - name: Run code coverage analysis
+        run: |
+          pipenv run python -m coverage run -m pytest
+          pipenv run python -m coverage report --include=*.py --fail-under=80
