+++ Pipfile
@@ -6,21 +6,27 @@ name = "pypi"
 [packages]
 pylint = "*"
 black = "*"
-flask = "==3.0.0"
+flask = "==2.0.1"
 requests = "*"
 crepe = "*"
 numpy = "*"
 scipy = "*"
 librosa = "*"
 pymongo = "==4.5.0"
-python-dotenv = "==0.16.0"
+python-dotenv = "*"
 flask-session = "*"
 pretty-midi = "*"
 pytest = "*"
-werkzeug = "==3.0.0"
+werkzeug = "==2.0.1"
 flask-cors = "==3.0.10"
 gunicorn = "==20.1.0"
 boto3 = "*"
+soundfile = "*"
+ffmpeg = "*"
+tensorflow = "*"
+pysoundfile = "*"
+update = "*"
+coverage = "*"
 
 [dev-packages]
