+++ machine_learning_client/dockerfile.ml
@@ -1,6 +1,6 @@
 FROM python:3.8
 RUN apt-get update && apt-get install -y ffmpeg
-WORKDIR /machine-learning-client
+WORKDIR /machine_learning_client
 COPY requirements.txt .
 RUN pip install --no-cache-dir -r requirements.txt
 RUN pip install git+https://github.com/librosa/librosa
