+++ pytest.ini
@@ -1,0 +1,3 @@
+[pytest]
+filterwarnings =
+    ignore:pkg_resources is deprecated:DeprecationWarning
